New Messagewindows comes up at every scan.
Now each new message is stored and identified by its id (msg.getMessageNumber()) 
In each scan new messages is checked if already read in and in that case they are ignored.
If user closes the message window and the message is still flaged as unread it will show a new window.